requirements.txt
astroid==4.0.4
dill==0.4.1
dnspython==2.8.0
flake8==7.3.0
isort==8.0.1
mccabe==0.7.0
platformdirs==4.9.4
psutil==7.2.2
pycodestyle==2.14.0
pyflakes==3.4.0
pylint==4.0.5
PyYAML==6.0.3
tomlkit==0.14.0
vulture==2.14
